[OPT]update to zy_compass_3.8:     1. add operators: GlobalPool/NonZero/SlotUpdate/QueryRebatch/Unique/IsInf/IsNaN     2. refine operators: BasicLSTM/GRU/Concat/Cumulate/Gather/NMS etc     3. add global calibration strategy: mvn_correction     4. enhance dynamic shape support for eltwise/reshape/batchnorm operator     5. add pass: eliminate_op, and refine pass: insert_op/merge_inserted_op     6. add opt_forward_main command tool     7. support qconfig plugin for easy quantization configuration, see Optimizer/plugins/aipubt_qconfig_bevformer.py for example

+++ AIPUBuilder/Optimizer/ops/convwinograd.py
@@ -434,9 +434,9 @@ def winograd_conv_1D_HP(self, inp, weights, bias, m=2, r=3, DEBUG=True):
         self.constants['WinogradWeights'] = winograd_weights_t
 
     if self.quantized:
-        Gg_weights = weights
+        Gg_weights = weights.float()
     else:
-        Gg_weights = self.attrs['WinogradWeights'].permute((0, 3, 1, 2))
+        Gg_weights = self.attrs['WinogradWeights'].permute((0, 3, 1, 2)).float()
 
     def DataReverseMethod(benchmark=False):
         data_reverse = torch.zeros((Batch, _Cin, LoopHeight, LoopWidth, _KernelHeight, TileInWidth), device=inp.device)
@@ -529,7 +529,7 @@ def winograd_conv_1D_HP(self, inp, weights, bias, m=2, r=3, DEBUG=True):
             _out = method3(benchmark)
         else:
             _out = method3(benchmark)
-        return _out
+        return _out.float()
 
     if DEBUG:
         StartTime = datetime.datetime.now()
